requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
django-filter==24.3
python-dotenv==1.1.1
Pillow==10.2.0
